requirements.txt
mlflow
dvc
fastapi
uvicorn[standard]
httpx
python-multipart
pydantic
pydantic-settings
scikit-learn
pandas
numpy
statsmodels
lightgbm
xgboost
prophet
pyarrow
tqdm
pyyaml
requests
pytest
pytest-cov
pre-commit
